Implement optional diff headers for empty diffs
This closes #18 by providing a configuration option that allows the
user to specify that in the case of a diff determining that two files
are equal, the diff headers should be displayed along with the text
that indicates there are no changes.

This can be useful in a variety of use cases; it allows you to detect
in a single file diff where the source file was located or to see the
timestamps of the related files in a bulk diff (e.g. for an expired
override) which is otherwise only displayed if there are changes.