Enhance report API to allow for view settings
For some upcoming changes, this adds to the ReportGenerationThread's
_set_content() method the ability to provide settings that will be
applied to the final output view when the report is finally displayed
to the user.

These settings are always applied to the output view; that is to say
they are applied after the call that creates or updates the report
view, to ensure that they are always updated, since output_to_view()
will only apply settings to newly created views.