Reports store the packages with expired overrides
Using the previously added API, all of the reports that display
override information now capture a list of all packages that have at
least one expired override.

The list is stored as a custom view setting, allowing commands to
easily determine what packages (if any) had expired overrides at the
time that the report was generated.

This is easier than trying to capture that information from the buffer
and less memory intensive than trying to persist the package
information after the report is generated.