Make ContextHelper expired override aware
This is a bit of a modification and cleanup to the code in the
ContextHelper class.

The first of the changes here is that the _extract() method now
assumes that the scope is a part of text.override-audit so that it
does not have to redundantly be provided in every extraction, which
makes things cleaner.

The _override_at_point() and view_context() methods now take a boolean
parameter named expired that tells them if the override that is pulled
from the buffer needs to be expired or not.

When True is provided, the exaction will only succeed if the override
at the cursor position is expired, while when the value is False it
works as it previously did and just requires an override at the cursor
position.

This allows context sensitive commands to optionally determine that
they will only operate on an expired override if they so desire.