Include threaded touch operation for overrides
Although this is not yet connected to any external commands yet, this
adds in a helper function and a background thread that can trigger a
freshen operation.

The operation can be carried out either for a single override file or
for all expired overrides in a given package; in the latter case care
is taken to only touch files that are considered to be expired as of
the time that the operation is started.

The console will indicate the names of all files that are freshened as
well as any that are not for some reason (does not exist, no
permission, etc) and the status bar and console will log a final
message indicating the result.

This message can indicate if the single override was touched or not,
or for a multiple operation it either tells you that all overrides
were touched or only some of them, allowing you to investigate
further.