Connect freshen command to packages in reports
This makes the necessary connections to provide the ability to bulk
freshen all of the expired overrides from within a single package.

The ContextHelper() class has been tweaked to provide an extra method
for determining if the current view contains the name of a given
package in its list of packages with expired overrides (which was
added in a765f1) so that commands can determine if they are applicable
or not.

In addition to this, the command for providing context menus on
package names has been extended to require an action parameter. The
possible values for this are "diff" (which the command did previously)
or "freshen".

The context menu command for this operation is offset from the
other OverrideAudit commands in the menu by a separator to keep it
visually distinct and help ensure that it doesn't get selected by
accident (although confirmation is enabled by default).