Connect freshen command to overrides in reports
This makes the necessary connections to provide a single file freshen
capability for expired overrides whose names appear in reports.

As for packages, the command is only visible to an override which is
marked as expired in the report. Unlike the report freshening code,
this does not currently modify the report to show the override as no
longer expired.

Support for that is currently planned to be a part of #20 when it is
implemented.

This should close #15, although it may be worth a new feature issue
for adding command palette commands and/or tab context commands for
freshening files (e.g. freshen on a single diff or freshen all from
a single package bulk diff).