Gracefully handle permission errors in diffs
While implementing #15 it was discovered that the code for collecting
the contents of an unpacked override did not gracefully handle errors
having to do with permissions being too restrictive to read the file.

Although it is highly unlikely that such a thing would actually occur,
it's not outside of the realm of possibility and it's easy enough to
more gracefully degrade to a console error message instead of a
complicated looking stack dump.