Remove expired marks on refreshed overrides
This partially implements #20 in that when expired overrides are
refreshed, the [X] mark on them that marks them as being expired is
removed from the report. This occurs both for a per-override refresh
or a refresh of an entire package at once.

Removing the mark from the override automatically causes the syntax to
change to indicate that the override is no longer expired, which has
the effect of removing the associated context items from that
override.

This partially implements #20 in that the marks on expired overrides
are modified. A future change will fully flesh out that issue, but
some more internal refactoring needs to happen first to keep
everything clean.