Update the PackageInfo __repr__ method
This makes the method generate output a little more user friendly, for
use in future debugging of the refactor.

Where the old code insisted on showing the found paths, we instead now
trust more than the paths found will be correct and display a shorter
version of that information that more closely resembles what
OverrideAudit would display in a report to make the output more
readable.