Refactor _deep_scan out of the PackageInfo class
This method was previously a static class member solely for use by
check_potential_override (and arguably should have been pulled out
when that was).

The new version is named _pkg_scan, and instead of always doing a
recursive search, it now takes a boolean parameter (defaulting to
False) to control this.

Most important, there was a filename case issue here, in that when
checking the directory listing to see if a file with the given name
was found, it would always check in a case sensitive manner even
though under Windows/MacOS that would cause it to not find a file if
the search was being conducted for a file with the wrong case.

As a result of this, the search is now done as a case insensitive
search as appropriate for the operating system in use.