Perform some cleanup in PackageList
This is just some miscellaneous cleanup of the code in the PackageList
class. Primarily it modifies all of the member variables that were
created after the initial code to use the same prefix as all of the
others.

This is purely cosmetic, but once I noticed that it was wrong, it
bothered me so I had to fix it.