Optimize override freshening
This change uses the new ability of PackageInfo to create a single
instance of itself to enhance override freshening so that instead of
always having to collect all information on all packages, it can
instead collect only information on the package that contains the
override(s) being freshened.

Along with this change the default value for the scan parameter to the
PackageInfo __init__ method is changed to True from False because
external code probably almost always mean for it to work this way. It
was originally the other way to maintain compatibility during the
refactor.