Create custom package lists with PackageList
This adds to the PackageList class the ability to pass a single
package name or a list of package names to influence the files and
folders used to construct the list.

When no arguments are given to the constructor, the full package list
is created (as it always has been), while a single name or list of
names will filter the list of found sublime-package files and unpacked
directories to just those who have names in the provided list.

On case insensitive file systems, the case of the package names passed
in the list doesn't matter and the ultimate package name used will be
taken from the files/folders found (as it works when there is no
filter).

The PackageListCollectionThread helper class has been modified to
accept an optional name_list parameter so that it's possible to use it
to collect only a subset of packages if desired.