Optimize Diff of single package or overrides
This uses the new functionality of the package list to constrain the
list of scanned packages to only a single package when performing a
bulk diff of a single package or when executing a diff of an override
from a context menu.

In both cases previously the entire list of packages would need to be
gathered even though it was known ahead of time that only a single
entry was needed.

In theory this could also be done by just creating a single
PackageInfo instance, but that would require the code to be modified
to be able to handle both a PackageList and a PackageInfo, so this
seems much cleaner as a solution.