Interface changes in packages.py
This makes two interface changes to packages.py.

Firstly, the _find_zip_entry module function has been renamed to
find_zip_entry so that it can be invoked from external code, since
things outside of the packages code may want to be able to look up the
contents of a package in a case insensitive manner.

Secondly, the PackageInfo class has sprouted a new package_mtime()
method, which returns the last modification time of the package file
that Sublime is currently using (installed if it exists, shipped if
not) or None if the package has no sublime-package files.

This goes hand in hand with the package_file() method which does the
same thing, returning the name of the package file instead.