Make OverrideFreshenThread future date aware
As a consequence of there being no time zone information on the time
stamps in a zip file, OverrideAudit treats all time stamps gained from
inside of a sublime-package file as being in the local timezone.

This raises a potential issue whereby based on the time zone
difference between your local machine and the machine of the person
who prepared the package, one or more entries inside of a package file
appear to be from the mysterious future until the local time advances
far enough.

In these cases, such overrides will appear to be expired because
regardless of what time the local file has, the packaged version of
the file is always newer. A secondary side effect of this is that
freshening such an override doesn't have any effect.

This commit modifies the freshening code so that it can detect when
the source of the file being freshened is from the future, and if so
it will touch the file to have a time one second farther in the future
than the source file has.