Include information on override freshening
This adds an extra section that talks about how to freshen an
override file by file or package by package. Currently this is only
listed in a section with no command listed, since the command in
question does not currently exist in the command palette.

This could conceivably be added as a set of commands that appear also
in the command palette for a report (possibly also being interactive?)
but until the BCR (Big Command Refactor) that is on hold.

Additional to the above, there is some extra clarification on the new
diff_empty_header configuration option and how it interacted with the
diff_unchanged setting.