Release version 1.1.1
A print statement for some debugging of one of the last bug fixes
prior to the release of 1.1.0 was accidentally left in and it upsets
my delicate sensibilities to have unadorned text sent to the console.

Additionally, I slightly bodged the release of 1.1.0 by carefully
crafting a messages file to be shown by PackageControl on package
update and then forgetting to update the control file to let PC know
that it existed.

Thus this small point release serves double duty by retroactively
fixing that problem and also providing a reason to display a new
message to anyone that might have had their package upgraded in the
interim.