Prepare for command refactoring
With the idea that the existing commands will be split apart to make
adding and maintaining them easier, as well as making key bindings and
context menus (and in future, phantoms/popups) easier to manage, this
gets the process started for that refactoring.

I had originally thought of having the helper code in a separate
helpers.py python file (added in the last commit) but upon more
reflection and diagram drawing, I think it's actually easier/better to
keep the override_audit.py file for that purpose (it was originally
going to eventually go away entirely).

To that end, all of the helper functions have had their names modified
to no longer have leading underscores, as they are now meant to be
used by outside code.