Factor context_override into separate commands
This factors the override_audit_context_override command into a set of
five commands in favor of having a single command that is responsible
for taking all actions based on a single "action" parameter.

This change is one of the main reasons for the BSR to take place in
the first place, as new commands that may be implemented would make
this single command solution exponentially harder to maintain as time
goes on.

As a side benefit of this, all of the commands have been modified so
that they can take "pkg_name" and "override" as parameters and be
executed by external code.

All of the existing context menus have been appropriately modified to
use the new commands, as have the default key bindings. Unfortunately
due to the changes, anyone with customized bindings will need to
modify them to the new commands.