Rename the command to bulk diff packages
The command that used to be named diff_package has been renamed to the
more appropriate diff_report, as it actually generates a bulk diff
report when you execute it, and all of the other report generation
commands use that name style.

Despite the name, the command is still meant to also be used to diff
an entire package, and so it is still invoked by the diff_single
command when you specify the bulk parameter.

In addition, some upcoming refactoring is going to split apart an
existing command to a name that might conflict with the name that this
command used to have.