Factor context_pacvkage into separate commands
Much like 4834cf, this factors the override_audit_context_package
command into a set of to commands in favor of having a single command
that is responsible for taking all actions based on a single "action"
parameter.

This command was less overloaded than the one for overrides, but it
still makes room for newer commands to be implemented in an easier
fashion.