Rename context_report to refresh_report
The last of the main part of the BSR, this renames the last command
that is responsible for refreshing a report and factors it out of the
core file and renames it to something that is more representative of
its actual function.

Also cleaned up is the nascent Changelog, which better represents all
of the command renames in a more succinct manner than previously.