Fix visibility regression in new package commands
As originally written, the commands were made available any time the
view context or command arguments contained a package. This however
caused the commands to also be visible via the context menu for
overrides as well because they also contain a package.

The fix is as simple as ensuring that there IS a package and that
there IS NOT an override.