Allow an extra form of package argument
Part of the set of imminent changes is modifications to all of the
commands to ensure that they are disabled when they cannot be executed
(primarily due to lack of appropriate arguments) and that it is
possible to invoke the commands originally meant for the context menu
by specifying arguments.

This change modifies the view_context() method of ContextHelper so
that it will accept "package" as an argument if "pkg_name" is not
provided, making hand written commands nicer looking.

This will become the standard in a future set of changes, so this is
just a stopgap for an unspecified period of time.