Ensure override marks are only a single character
The current intention is that the marks on overrides in packages (i.e.
the text between an optional set of `[ ]` prefixed on the name) would
be only a single character in length.

This modifies the command to enforce this particular restriction.

Additionally this command is the first one to be swapped to use
"package" as an argument instead of "pkg_name".