Ensure that report refresh works only for reports
The command for refreshing a report now contains an is_enabled handler
to disable the command when the view that it is being invoked on is
not a report.

This stops errors from being displayed if the command is invoked
without the appropriate context in place (say from a menu item).