Modify context commands to take arguments
This modifies the five new commands that are currently used only on
the context menu for overrides (edit, diff, delete, toggle, freshen)
so that it is possible to invoke them directly via command arguments
as well as via contextual information.

All commands will also block themselves from executing when they do
not have the proper information to execute.