Fix the last usages of pkg_name in commands
This is (I think) all of the cases of using pkg_name as an argument to
a Command to specify the package. This also tweaks the diff_override
command, for which I forgot to make the swap for the local variables.

There are other API functions (all internal) which use pkg_name
instead, but those will probably remain as is.

Additionally, the tweak to allow both "pkg_name" and "package" as
arguments remains; this will be removed (if I remember) at the next
full test phase, probably prior to the release of the current dev
branch so as to ensure that the BSR didn't destroy any functionality.