Start of Context refactoring; include helper tuple
This change is a modification to the ContextHelper base class and the
commands that use it to collect their context information (i.e. the
package and override to operate on) to use a custom tuple instead of
a regular one.q

There is now a CommandContext class which is essentially a named tuple
but with some extra methods that make the code doing tests on contexts
a little cleaner. This is implemented in a manner similar to how
collections.namedtuple() would create the class.

This new tuple has an extra field in it that indicates where the
context information came from (i.e. settings, arguments or from
examining the buffer at the mouse location).

Now the ContextHelper.view_context() method returns one of these
tuples. Since the length of the tuple has changed, all commands that
invoke this method have been somewhat refactored to use the new method
which in most cases makes them much shorter.

There is still some work to be done in this area (adding extra helpers
to the CommandContext class) but this is a necessary first step to
some upcoming commits that will alter how the commands render
themselves visible or not in the context menu.