Enhance ContextHelper for upcoming commits
This adds a couple of new helper methods to the class in preparation
for upcoming changes.

One of the goals is to have commands always remain visible in menus
unless you specifically give them the right to hide themselves when
they do not currently apply, so that users can customize menus and get
the intended interactions.

To this end, there is now a method that checks for an optional
argument to see if the command should always be visible, which
defaults to True if not given. This can be checked by commands to
short circuit their own tests and just always remain visible.

Additionally, since many commands should not be active if the override
they are going to operate on no longer exists, there is now a helper
method that checks not only that a command context has a package and
override, but also that the combination is actually a file that
currently exists.