Update command visibility, part one
This modifies the commands to edit, diff, delete and toggle overrides
and the command to refresh reports so that they have the following
properties:

- Always remain visible unless an optional parameter allows them to
  hide themselves when they do not apply

- Captions for commands that operate on an override do not include the
  name of the override they will operate on if the command is being
  executed directly from an edit or diff buffer; instead they just say
  "this override"

- Commands that operate on overrides will only be enabled if the
  underlying override file actually exists (along with the previous
  tests for eligibility)

- Command captions for disabled commands that are disabled because
  the appropriate arguments/context does not exist are shortened to
  just say what they would do if they were enabled without having
  e.g. the name of an override in them (e.g. "Diff Override" instead
  of "Diff Override 'None'")

All commands that will be touched now have their entries in the
context menus have the new `always_visible` argument, which is set to
`false` for all commands that have been updated so far so that they
operate as they did and only appear when they should. The rest of the
commands have the value set to `true` although they still remain
hidden as they have not been touched yet.