Include an override diff/edit swap in menu
This adds in a command to `Tools > Override Audit` that swaps between
a diff of an override and its edit session.

Previously this was only available via the command palette and via
key bindings.

The command disables itself when the current view is not a diff or
edit of an override.