Fix syntax issue introduced at 745f8c in #24
While making this modification to ensure that overrides that start
with a period (e.g. .gitignore) are properly identified, a bug was
introduced in which the text in override and diff reports that tells
you that there are no overrides was incorrectly identified as an
override itself.

The syntax test did not catch this due to it checking only for a base
scope and not for the exclusion of the override scope as well.

In order to resolve this, the syntax has been updated and the text in
question is now wrapped in < > marks instead of [ ] since the former
is not valid in a filename but the latter is.