Backport the `save_on_diff` fix from v2.0.0
Back when we were doing BSR to split out the code nicely for version
2.0.0, we fixed a bug in the `save_on_diff` handling wherein if you
have the `trim_trailing_white_space_on_save` turned on, a line has any
trailing white space AND the file is unsaved, the file will be saved to
disk before doing the diff, but it will be left marked as dirty.

The fix here is similar to the one created for version 2, but the code
is a little more simplistic here.