Remove some backported fixes from the changelog
Some fixes that were originally made in the version 2.0.0 branch as a
part of the BSR have been backported to the legacy branch, where they
have been added to the changelog and will be a part of the next release.

As such, these changes no longer need to be in the changelog here, since
that would look a little weird and wrong.