Detect locally installed dependencies properly
This fixes #25 by properly detecting that an unpackaged package is a
locally installed dependency.

Normal dependencies are installed by PackageControl and have a
dependency-metadata.json file created for them in their package
directory when they are installed.

While developing a new dependency, this file does not exist yet.
However in order to have the package properly detected as a dependency
PackageControl allows you to create a .sublime-dependency file to
indicate the appropriate load order.

This commit modifies the behaviour for checking if an unpacked package
is a dependency by checking for the presence of that file if the
normal one is not there.