Swap opening a README file for online docs
This completes a change that could have been done a ways back when the
final official version of the online documentation was completed, and
swaps the command palette and main menu entries for opening the readme
file from the package for commands that open the documentation in the
user's browser instead.

In some future build, this can be swapped for in-Sublime help via
hyperhelp.