Release verison 1.1.2
This includes a fix for the `save_on_diff` functionality, repairs a
regression in the syntax added in version 1.1.0, and replaces the menu
and command palette entries that used to display the README file for new
ones that open the online documentation in the user's browser.

This is an interim point release while we work some more on the external
diff functionality that will be in the next version, in preparation for
our abandonment of the 1.x versions in favour of the new 2.x BSR version
of the code.