Include a new setting, `external_diff`
This adds a new setting to the system to specify the name of the
external diff program to use on overrides (if any), and how to execute
it.

This is planned to be done in a way that's extensible in the future and
which also supports per-platform settings in the single settings entry
so that people that use Sublime on multiple platforms (such as myself)
can "set it and forget it" and not have to keep fiddling with settings.

whoops