Include a new PackageInfo API endpoints
This includes new API endpoints in the PackageInfo class to retrieve the
contents of a packed or unpacked override as a list of lines in the file
as well as (for packed overrides) an indication as to whether the source
is a shipped or installed package.

This is a thin wrapper around existing API endpoints that are internal
and are used in internal diffing.

whoops