Include an internal helper for override extraction
This includes a simple helper function whose job it is to take the info
on a package and the name of an override, and extract the appropriate
packed file out into a temporary file so that it can be examined by
tools outside of Sublime.

The name of the file created is returned, and consists of information
that tells you if the file is shipped or installed, and the full package
path and override name (plus an unfortunate 6 character uniqueness
suffix on the name proper).

The file is created as read only to ensure that the code using the file
is made aware that any changes will be lost, since the process is going
to remove the file once the diff is completed.