Implement the initial external diff open command
This does the work of implementing the ability to open a diff in an
external program, assuming that one is configured.

Currently this is only available from the context menu and command
palette from inside of an existing diff window; i.e. there is no direct
way to diff externally without opening an internal diff first.

This may change either before or after final release, but this is an
attempt to keep the clutter in the context menu down by not having the
option available in a variety of circumstances since in many cases a 
more simple diff is probably enough.