Add a new thread class for External Diffing
This includes a new thread class for performing an external diff by
extracting the base file out of a `sublime-package` file into a temp
file (using a recent change) and then executing an external program.

This thread silently waits for the process to finish and then attempts
to remove the temporary file if it can, although it does not complain if
it cannot.

This is the core of functionality for opening a diff in an external
program.