Constrain variable expansions for external_diff
The previous processing for the external diff setting was to expand all
variables across the whole of the argument dictionary, which is not the
right thing to do.

In particular, Sublime only expands variables in `shell_cmd`, `cmd` and
`working_dir`. Additionally, the `env` and `path` keys may also have
variables in them, but they are given to the OS to expand out based on
their being environment variables.

As such, if you try to include `$path` in the `path` key to include the
original path, our variable expansion will clobber it out since unknown
variables expand as empty strings. The same can happen to `env`.

To get around that, we now explicitly only expand variables in the
`shell_cmd` and `working_dir` arguments (and in the future, `cmd`, if we
decide to support it).