Remove our handling and support for a direct path
As mentioned in a recent stream, there is a bug in the core exec code
regarding the `path` argument in a build system which I was unable to
remember how to replicate at the time.

The gist of the bug is that if a build system contains a `path` argument
the `AsyncProcess` class will save the current path variable and then
modify the process environment with the provided path. It will then spin
up the child process and reset the path back to what it was.

If the call to `Popen` fails (for example with a `cmd` that specifies an
invalid program), the exception thrown causes the code to skip over the
reset of the environment.

Thus in this case the environment is left in a modified state, which
affects all future builds. If a person used an invalid or short path
(for example to specify just a single folder), the path is forever
altered.

Although the fix for this is easy enough, the `path` argument is not one
of the documented keys in a build system and may be a hold over from a
previous Sublime version.

Since the same thing can be accomplished with `env`, here we are just
redacting the `path` handling code entirely instead.