Release version 1.2.0
This introduces the ability to run an external diff program in order to
see the results of an override diff in the tool of your choice.

This is controlled via a setting by which the user can specify exactly
what program they want to execute in order to do the diff and what the
command line for that tool should look like.

At the present time, this functionality is in the command palette and
the context menu of diff views with wording that indicates that you can
open this diff in an external tool.

A future version might update this to make the original diff open in
that tool or something similar.