Make the license copyright dates a range
Just a small tweak, since I recently updated the copyright notice on the
license to the latest date, I've decided that I'd like this to be a
range of dates instead.

That conveys the idea that the code has been around for a while and is
thus mature(ish) while also showing that it's still actively maintained.