Generate a dialog on external diff failure
This generates a dialog box that makes it apparent when an external diff
operation has failed, where failure is an indication that the program
that was launched returned a non-zero exit code.

The most obvious cause of this is a problem with the `shell_cmd` using a
program that cannot be found on the path.