Improve detection of invalid external_diff setting
This makes some additional changes to the code that determines if we
should allow an external diff or not by checking a few common errors.

The most important of these is that it used to be possible to specify
a value of `true` here, which would generate an error because the code
would consider the value a dictionary.

Now any boolean value always disables the external diff functionality,
and the only thing that enables it is a non-empty dictionary or a
string of "sublimerge".

Existing logic generates errors if the dictionary does not have the
right keys when the diff happens, which saves us the time of having to
try and valid the dictionary every time the menu is displayed.