Release version 1.2.1
This extends the previous functionality for external diffs by allowing
the Sublimerge Pro or Sublimerge 3 package to be used to perform the
external diff action by setting the `external_diff` setting to the
string "sublimerge".

This detects if the package is actually installed and only enables the
command when it is believed that the package is ready to go. This
includes detection of whether or not the package is currently ignored.

There is also a tweak to the detection of external diffing being allowed
that ensures that the command is not accidentally visible if the setting
of the `external_diff` setting is invalid.