Move all resources into their own part of the tree
In order to keep the top level of the package cleaner, this change
creates a new `resources` folder and puts all of the resources inside of
it in subfolders.

Along the way we need to modify the menus and the command palette so
that the entries for opening settings and key bindings still work.

Similarly, the code needed to be updated to know where to find the
syntax resources and the syntax tests needed to be similarly updated.