Visually refresh the package report table frame
This simple modification adds `+` characters to the header and footer
lines in the package report table, which is an overall better look
w/regards to the "ASCII art" nature of the display.

I spent some time digging and trying to determine why I didn't do this
in the first place, and my best guess is either laziness or something
that was just held over from the initial quick prototypes.